+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 
 ### Changed
 
+- allow get and put as `one=True` single-item `Dict` or `MatchMapping` instead of error, makes more sense
 - simplified `trivia` option to always specify entire state, no more merging with current state
 
 ### Fixed